Issue #6728 - Fixes for QPACK

+++ jetty-http3/http3-qpack/src/test/java/org/eclipse/jetty/http3/qpack/EncodeDecodeTest.java
@@ -75,7 +75,7 @@ public class EncodeDecodeTest
     public void test() throws Exception
     {
         // B.1. Literal Field Line With Name Reference.
-        int streamId = 0;
+        long streamId = 0;
         HttpFields httpFields = HttpFields.build().add(":path", "/index.html");
 
         ByteBuffer buffer = BufferUtil.allocate(1024);
